Extract exception classes to optuna.exceptions module.
