+++ optuna/integration/chainer.py
@@ -1,6 +1,7 @@
 from __future__ import absolute_import
 
 import optuna
+import optuna.exceptions
 from optuna import type_checking
 
 try:
@@ -106,7 +107,7 @@ class ChainerPruningExtension(Extension):
         self.trial.report(current_score, step=current_step)
         if self.trial.should_prune():
             message = "Trial was pruned at {} {}.".format(self.pruner_trigger.unit, current_step)
-            raise optuna.structs.TrialPruned(message)
+            raise optuna.exceptions.TrialPruned(message)
 
 
 def _check_chainer_availability():
