+++ optuna/integration/chainermn.py
@@ -2,10 +2,10 @@ from __future__ import absolute_import
 
 import gc
 
+from optuna.exceptions import TrialPruned
 from optuna.logging import get_logger
 from optuna.storages import InMemoryStorage
 from optuna.storages import RDBStorage
-from optuna.structs import TrialPruned
 from optuna.trial import BaseTrial
 from optuna import type_checking
 import warnings
