+++ optuna/integration/pytorch_ignite.py
@@ -1,4 +1,5 @@
 import optuna
+import optuna.exceptions
 from optuna import type_checking
 
 if type_checking.TYPE_CHECKING:
@@ -57,7 +58,7 @@ class PyTorchIgnitePruningHandler(object):
         self.trial.report(score, engine.state.epoch)
         if self.trial.should_prune():
             message = "Trial was pruned at {} epoch.".format(engine.state.epoch)
-            raise optuna.structs.TrialPruned(message)
+            raise optuna.exceptions.TrialPruned(message)
 
 
 def _check_pytorch_ignite_availability():
