+++ optuna/integration/tfkeras.py
@@ -1,6 +1,7 @@
 from __future__ import absolute_import
 
 import optuna
+import optuna.exceptions
 from optuna import type_checking
 
 if type_checking.TYPE_CHECKING:
@@ -64,7 +65,7 @@ class TFKerasPruningCallback(Callback):
         # Prune trial if needed
         if self.trial.should_prune():
             message = "Trial was pruned at epoch {}.".format(epoch)
-            raise optuna.structs.TrialPruned(message)
+            raise optuna.exceptions.TrialPruned(message)
 
 
 def _check_tensorflow_availability():
