+++ tests/integration_tests/test_tfkeras.py
@@ -4,6 +4,7 @@ import pytest
 import tensorflow as tf
 
 import optuna
+import optuna.exceptions
 from optuna.integration import TFKerasPruningCallback
 from optuna.testing.integration import create_running_trial
 from optuna.testing.integration import DeterministicPruner
@@ -52,8 +53,8 @@ def test_tfkeras_pruning_callback_observation_isnan():
     trial = create_running_trial(study, 1.0)
     callback = TFKerasPruningCallback(trial, 'loss')
 
-    with pytest.raises(optuna.structs.TrialPruned):
+    with pytest.raises(optuna.exceptions.TrialPruned):
         callback.on_epoch_end(0, {'loss': 1.0})
 
-    with pytest.raises(optuna.structs.TrialPruned):
+    with pytest.raises(optuna.exceptions.TrialPruned):
         callback.on_epoch_end(0, {'loss': float('nan')})
