+++ tests/samplers_tests/tpe_tests/test_sampler.py
@@ -1,8 +1,8 @@
 import optuna
 
+from optuna.exceptions import TrialPruned
 from optuna.samplers import tpe
 from optuna.samplers import TPESampler
-from optuna.structs import TrialPruned
 
 if optuna.type_checking.TYPE_CHECKING:
     from optuna.trial import Trial  # NOQA
